requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
jinja2
pydantic
google-generativeai
resend
